+++ src/message/address.cpp
@@ -143,7 +143,7 @@ void address::to_data(uint32_t version, writer& sink) const
         net_address.to_data(version, sink, true);
 }
 
-uint64_t address::serialized_size(uint32_t version) const
+size_t address::serialized_size(uint32_t version) const
 {
     return variable_uint_size(addresses_.size()) +
         (addresses_.size() * network_address::satoshi_fixed_size(version, true));
